tornado_probability.py
"""
Tornado Probability Map Generator
==================================
Fetches HRRR forecast data for hours 1-6 and computes a composite tornado
probability index from:
  - Convective Available Potential Energy (CAPE)
  - 0-6 km bulk wind shear
  - Significant Tornado Parameter (SigTor)
  - 1-hr max 0-3 km AGL Updraft Helicity (UH)

Probability is estimated for a 25-mile radius influence zone and exported as
an interactive Folium HTML map centred on the Northeast US.
"""

import argparse
import json
import os
import sys
import warnings
from datetime import datetime, timezone
from pathlib import Path

import numpy as np

warnings.filterwarnings("ignore")

# ---------------------------------------------------------------------------
# Northeast US bounding box  (lat_min, lat_max, lon_min, lon_max)
# ---------------------------------------------------------------------------
NE_BBOX = (37.0, 47.5, -83.5, -66.0)

# ---------------------------------------------------------------------------
# Thresholds used to normalise each parameter (values at/above → score = 1)
# ---------------------------------------------------------------------------
CAPE_MAX = 3000.0       # J kg-1   (extreme instability)
SHEAR_MAX = 40.0        # m s-1    (extreme deep-layer shear)
SIGTOR_MAX = 4.0        # unitless (very high significant-tornado parameter)
UH_MAX = 200.0          # m² s-2   (very large updraft helicity)

# Weights that sum to 1 – UH and SigTor carry the most skill
WEIGHTS = dict(cape=0.20, shear=0.20, sigtor=0.30, uh=0.30)


# ---------------------------------------------------------------------------
# Helpers
# ---------------------------------------------------------------------------

def _clamp01(arr: np.ndarray) -> np.ndarray:
    """Clamp values to [0, 1]."""
    return np.clip(arr, 0.0, 1.0)


def _logistic(x: np.ndarray, k: float = 8.0, x0: float = 0.35) -> np.ndarray:
    """Smooth logistic squeeze so moderate composites yield moderate probs."""
    return 1.0 / (1.0 + np.exp(-k * (x - x0)))


def composite_to_probability(cape_norm, shear_norm, sigtor_norm, uh_norm):
    """
    Combine normalised fields into a tornado probability [0, 1].

    Steps:
      1. Weighted average composite index  → [0, 1]
      2. Require CAPE > 0 (no instability → no tornado)
      3. Apply logistic function to spread the distribution
      4. Scale to realistic tornado-probability range (max ~40 %)
    """
    composite = (
        WEIGHTS["cape"] * cape_norm
        + WEIGHTS["shear"] * shear_norm
        + WEIGHTS["sigtor"] * sigtor_norm
        + WEIGHTS["uh"] * uh_norm
    )
    # Suppress where there is essentially no CAPE
    composite = np.where(cape_norm < 0.01, 0.0, composite)
    prob = _logistic(composite)
    # Rescale: logistic output lives in (0,1); map so composite=1 → ~40 %
    prob = prob * 0.40
    return _clamp01(prob)


# ---------------------------------------------------------------------------
# HRRR fetching
# ---------------------------------------------------------------------------

def fetch_hrrr_fields(run_time: datetime, fxx: int):
    """
    Return a dict of 2-D numpy arrays for the NE bounding box at forecast
    hour *fxx*.  Fields: lat, lon, cape, shear, sigtor, uh.

    Falls back to realistic synthetic data if Herbie / the HRRR archive is
    unavailable (useful for CI, GitHub Actions, or offline demos).
    """
    try:
        from herbie import Herbie  # type: ignore

        print(f"  Fetching HRRR fxx={fxx:02d} …", flush=True)

        H = Herbie(run_time, model="hrrr", product="sfc", fxx=fxx)

        # --- CAPE (surface-based) -------------------------------------------
        ds_cape = H.xarray("CAPE:surface")
        cape_var = [v for v in ds_cape.data_vars if "cape" in v.lower() or "CAPE" in v][0]
        cape_da = ds_cape[cape_var]

        # --- 0-6 km bulk shear (use VUCSH / VVCSH or compute from winds) ----
        try:
            ds_shear = H.xarray(":VUCSH:0-6 km above ground")
            ds_shearv = H.xarray(":VVCSH:0-6 km above ground")
            ushear_var = list(ds_shear.data_vars)[0]
            vshear_var = list(ds_shearv.data_vars)[0]
            shear_da = np.sqrt(
                ds_shear[ushear_var].values ** 2
                + ds_shearv[vshear_var].values ** 2
            )
            lats = ds_shear["latitude"].values
            lons = ds_shear["longitude"].values
        except Exception:
            # Fallback: estimate shear magnitude from cape grid coords only
            lats = cape_da["latitude"].values
            lons = cape_da["longitude"].values
            shear_da = np.zeros_like(lats)

        # --- Significant Tornado Parameter ----------------------------------
        try:
            ds_stp = H.xarray(":FRICV:surface")   # STP stored as FRICV
            stp_var = list(ds_stp.data_vars)[0]
            stp_da = ds_stp[stp_var].values
        except Exception:
            stp_da = np.zeros_like(lats)

        # --- 1-hr max 0-3 km AGL Updraft Helicity ---------------------------
        try:
            ds_uh = H.xarray(":MXUPHL:0-3000 m above ground")
            uh_var = list(ds_uh.data_vars)[0]
            uh_da = ds_uh[uh_var].values
        except Exception:
            uh_da = np.zeros_like(lats)

        cape_arr = cape_da.values
        if isinstance(shear_da, np.ndarray) is False:
            shear_da = shear_da.values

        # Subset to NE bounding box -----------------------------------------
        lat_min, lat_max, lon_min, lon_max = NE_BBOX
        # HRRR uses 0-360 lon; convert bbox
        lon_min360 = lon_min % 360
        lon_max360 = lon_max % 360

        mask = (
            (lats >= lat_min) & (lats <= lat_max)
            & (lons >= lon_min360) & (lons <= lon_max360)
        )
        if mask.sum() == 0:
            # Try with negative lons
            mask = (
                (lats >= lat_min) & (lats <= lat_max)
                & (lons >= lon_min) & (lons <= lon_max)
            )

        lat_sub = lats[mask]
        lon_sub = lons[mask]
        # Convert 0-360 lons back to -180 to 180
        lon_sub = np.where(lon_sub > 180, lon_sub - 360, lon_sub)

        return dict(
            lat=lat_sub,
            lon=lon_sub,
            cape=cape_arr[mask],
            shear=shear_da[mask] if shear_da.shape == lats.shape else np.zeros(mask.sum()),
            sigtor=stp_da[mask] if stp_da.shape == lats.shape else np.zeros(mask.sum()),
            uh=uh_da[mask] if uh_da.shape == lats.shape else np.zeros(mask.sum()),
        )

    except Exception as exc:
        print(f"  [WARN] Could not fetch live HRRR data (fxx={fxx}): {exc}")
        print("  Using synthetic demonstration data for this forecast hour.")
        return _synthetic_fields(fxx)


def _synthetic_fields(fxx: int) -> dict:
    """
    Generate physically plausible synthetic HRRR-like fields for the NE US.
    Used when live data is unavailable.  Values evolve realistically with fxx.
    """
    rng = np.random.default_rng(seed=42 + fxx)

    lat_min, lat_max, lon_min, lon_max = NE_BBOX
    # ~3-km HRRR-resolution grid over the NE
    n_lat, n_lon = 120, 180
    lats_1d = np.linspace(lat_min, lat_max, n_lat)
    lons_1d = np.linspace(lon_min, lon_max, n_lon)
    lon2d, lat2d = np.meshgrid(lons_1d, lats_1d)

    # ---- Gaussian "storm cluster" that drifts ENE with time ----------------
    # Cluster 1 – inland mid-Atlantic
    c1_lat = 39.5 + fxx * 0.15
    c1_lon = -77.0 + fxx * 0.20
    # Cluster 2 – southern New England
    c2_lat = 41.8 + fxx * 0.10
    c2_lon = -72.5 + fxx * 0.25

    def gaussian(clat, clon, sigma_lat=1.5, sigma_lon=2.0):
        return np.exp(
            -((lat2d - clat) ** 2) / (2 * sigma_lat ** 2)
            - ((lon2d - clon) ** 2) / (2 * sigma_lon ** 2)
        )

    envelope = np.clip(gaussian(c1_lat, c1_lon) + 0.6 * gaussian(c2_lat, c2_lon), 0, 1)

    # Add small-scale noise
    noise = rng.uniform(0, 0.15, envelope.shape)
    envelope = np.clip(envelope + noise, 0, 1)

    cape = envelope * 2800 * (0.7 + 0.3 * rng.uniform(size=envelope.shape))
    shear = envelope * 38 * (0.6 + 0.4 * rng.uniform(size=envelope.shape))
    sigtor = envelope * 3.5 * (0.5 + 0.5 * rng.uniform(size=envelope.shape))
    uh = envelope * 190 * (0.4 + 0.6 * rng.uniform(size=envelope.shape))

    return dict(
        lat=lat2d.ravel(),
        lon=lon2d.ravel(),
        cape=cape.ravel(),
        shear=shear.ravel(),
        sigtor=sigtor.ravel(),
        uh=uh.ravel(),
    )


# ---------------------------------------------------------------------------
# Main computation
# ---------------------------------------------------------------------------

def compute_probability(fields: dict) -> np.ndarray:
    cape_norm = _clamp01(fields["cape"] / CAPE_MAX)
    shear_norm = _clamp01(fields["shear"] / SHEAR_MAX)
    sigtor_norm = _clamp01(fields["sigtor"] / SIGTOR_MAX)
    uh_norm = _clamp01(fields["uh"] / UH_MAX)
    return composite_to_probability(cape_norm, shear_norm, sigtor_norm, uh_norm)


def apply_25mile_smoothing(lats, lons, probs):
    """
    Smooth raw probabilities so each point reflects the max probability
    within a ~25-mile (40 km) radius.  Uses a simple distance-weighted
    maximum via a vectorised nearest-neighbour approach for performance.
    """
    from scipy.spatial import cKDTree  # type: ignore

    # Convert to radians for haversine-compatible tree
    coords_rad = np.deg2rad(np.column_stack([lats, lons]))
    tree = cKDTree(coords_rad)

    # 25 miles ≈ 40 km; Earth radius ≈ 6371 km → angular radius in radians
    radius_rad = 40.0 / 6371.0

    smoothed = np.zeros_like(probs)
    for i, (pt, p) in enumerate(zip(coords_rad, probs)):
        idxs = tree.query_ball_point(pt, radius_rad)
        smoothed[i] = probs[idxs].max() if idxs else p

    return smoothed


# ---------------------------------------------------------------------------
# Entry point (called from map generator)
# ---------------------------------------------------------------------------

def run(run_time: datetime, output_dir: Path, fxx_list=None):
    """
    Process all requested forecast hours and return a list of result dicts.

    Each result dict contains: fxx, valid_time, lat, lon, probability.
    """
    if fxx_list is None:
        fxx_list = list(range(1, 7))

    output_dir.mkdir(parents=True, exist_ok=True)
    results = []

    for fxx in fxx_list:
        print(f"\n[Hour +{fxx:02d}]")
        fields = fetch_hrrr_fields(run_time, fxx)
        probs = compute_probability(fields)

        print(f"  Applying 25-mile smoothing …", flush=True)
        probs_smooth = apply_25mile_smoothing(fields["lat"], fields["lon"], probs)

        valid_dt = run_time.replace(tzinfo=timezone.utc)

        results.append(
            dict(
                fxx=fxx,
                valid_time=valid_dt.isoformat(),
                lat=fields["lat"],
                lon=fields["lon"],
                probability=probs_smooth,
            )
        )

    return results


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Compute HRRR tornado probabilities")
    parser.add_argument(
        "--run-time",
        default=datetime.utcnow().strftime("%Y-%m-%d %H:00"),
        help="Model run time (UTC), e.g. '2024-05-15 18:00'",
    )
    parser.add_argument(
        "--output-dir", default="maps", help="Directory for output files"
    )
    args = parser.parse_args()

    rt = datetime.strptime(args.run_time, "%Y-%m-%d %H:%M")
    run(rt, Path(args.output_dir))
